add "fileNamePostfix" to annotateMouseModulesGO
In the following function:

annotateMouseModulesGO = function (colorsCoexpr, transcriptInfoMouse, type, fileNamePostfix){...}

fileNamePostFix has been added so that the user may create separate files for different comparison groups in
the same directory; otherwise, if this function is run again for another group, it will overwrite the previously
generated file (e.g. when you run the function for males and females separately pointing to the same directory).Therefore,
the user can specify a postfix that identifies the group context in each call to this function.